AddProperty modal separated out into a separate component - all tested and functioning as it should be. add & all deleted as surplus files only used to hold redundant code.
