--- estate-agent-website/src/components/AddProperty.jsx
@@ -1,16 +1,0 @@
-// import { useState } from "react";
-// import "../CSS/AddProperty.css";
-// import { SiThenorthface } from "react-icons/si";
-
-// import DisplayProperty from "./DisplayProperty";
-
-// export default function AddProperty({fetchData}) {
-//   // create state
-
-
-  
-
-//   return (
-//     <p>blank</p>
-//   );
-// }
